requirements.txt
# ---- Core (recommended) ----
numpy>=1.23
scipy>=1.10
pandas>=2.0
PyYAML>=6.0
tqdm>=4.67
requests>=2.31
pillow>=10.0
matplotlib>=3.7

# ---- CV / Augmentation ----
opencv-python>=4.8
scikit-image>=0.21
scikit-learn>=1.3
albumentations>=1.4

# ---- PyTorch (install per CUDA in README if needed) ----
torch>=2.4
torchvision>=0.19
torchaudio>=2.4

# ---- Nice-to-have ----
rich>=13.9
seaborn>=0.13
